// front-end/types/actor-type.ts
export interface ActorType {
  _id: string;
  actor_id: number;
  name: string;
  gender: number;
  original_name: string;
  adult: boolean;
  known_for_department: string;
  profile_path: string;
  gender_name: string;
  also_known_as: string[];
  __v: number;
  createdAt: string;
  updatedAt: string;
}